clean & norm line_edition/*.c
