+++ includes/line_edition.h
@@ -6,7 +6,7 @@
 /*   By: cyfermie <cyfermie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/22 15:45:45 by cyfermie          #+#    #+#             */
-/*   Updated: 2018/11/09 17:13:34 by cyfermie         ###   ########.fr       */
+/*   Updated: 2018/11/27 16:24:46 by cyfermie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -362,12 +362,14 @@ int				write_one_char(int c);
 */
 struct s_line	*access_le_main_datas(void);
 void			add_history(const char *input, struct s_line *le);
+t_kno			get_key_number(const char *key);
 void    		handle_window_resize(struct s_line *le);
 void			le_free_datas(void);
 void			le_free_history(struct s_line *le);
 char			*line_edition(int prompt_type);
 t_kno			get_key_number(const char *key);
 void			process_key(struct s_line *le);
+char			*read_key(char key[LE_KEY_BUFFER_SIZE], struct sigaction *le_sig);
