+++ srcs/line_edition/access_le_main_datas.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   access_le_main_datas.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
+/*   By: cyfermie <cyfermie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/28 16:14:46 by cfermier          #+#    #+#             */
-/*   Updated: 2018/09/02 20:51:42 by jjaniec          ###   ########.fr       */
+/*   Updated: 2018/11/27 15:47:35 by cyfermie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
